Desktop/clips-as-server/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
exceptiongroup==1.2.2
faiss-cpu>=1.7.0
fastapi>=0.68.0
filelock==3.18.0
flake8==7.2.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio>=1.48.2
h11==0.16.0
huggingface-hub==0.30.2
identify==2.6.10
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isort==6.0.1
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.2.1
nodeenv==1.9.1
numpy>=1.19.2
packaging==25.0
pathspec==0.12.1
Pillow>=9.0.0
platformdirs==4.3.7
pre_commit==4.2.0
protobuf>=3.19.6,<7.0.0
pycodestyle==2.13.0
pydantic>=2.0.0
pydantic_core==2.33.1
pyflakes==3.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tensorboard>=2.12.0
tensorboard-data-server==0.7.2
tokenizers==0.21.1
tomli==2.2.1
torch>=2.0.0
torchvision>=0.15.0
tqdm==4.67.1
transformers>=4.30.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3>=1.26.14,<2.0.0
uvicorn>=0.15.0
virtualenv==20.30.0
Werkzeug==3.1.3
zipp==3.21.0
redis>=5.0.0
qdrant-client>=1.6.0
aiohttp>=3.8.0

# LLaVA-specific dependencies
accelerate>=0.20.3
sentencepiece>=0.1.99
bitsandbytes>=0.41.0
einops>=0.6.1
flash-attn>=2.3.0

# SigLIP/CLIP models
open_clip_torch>=2.20.0
transformers>=4.24.0

# Dataset handling
datasets>=2.4.0
huggingface_hub>=0.10.0

# Reranking
sentence-transformers>=2.2.2

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.12.0

# API
fastapi>=0.80.0
uvicorn>=0.17.0
pydantic>=1.9.0

# Testing
pytest>=6.2.5
hypothesis>=6.14.0

# Qwen-VL specific dependencies
transformers>=4.36.0
modelscope>=1.9.5
cpm_kernels>=1.0.11
timm>=0.6.13
gradio>=3.50.2
mdtex2html
einops>=0.6.1
tiktoken>=0.5.1 